appIcons: Give urgent windows more priority on activation
If urgent windows are available activation action should give priority
to them so that they can be focused and eventually unmarked as urgent.

So make getInterestingWindows() to list such windows as first ones, and
adapt activation code to ignore the focused state in case urgent windows
are there, prioritizing focusing them.